requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
drf-spectacular==0.29.0
openpyxl==3.1.5
python-dotenv==1.2.2
phonenumbers==9.0.27
psycopg[binary]==3.3.3
